+++ src/cli/cli.ts
@@ -6,20 +6,20 @@
  * The CLI parses command-line arguments, validates them, and executes
  * the corresponding command logic.
  */
-
-import version from "../version";
+import { version, help } from "./commands";
 
 const [...args] = process.argv.slice(3);
 
 switch (args[0]) {
-    case "-v":
-        console.log(version);
+    case "version":
+        version();
         break;
 
     case "help":
+        help();
         break;
     default:
-        console.log(`Unknown command: "${args[0]}"
+        console.log(`Unknown command: "${args[0] ?? "No command entered"}"
 
         To see a list of supported databridge commands, run:
         databridge help`)