+++ src/cli/commands/index.ts
@@ -1,0 +1,7 @@
+import version from "./version";
+import help from "./help";
+
+export {
+    version,
+    help,
+}